Shutting up lint errors
Current implementation uses stuff directly from androidx libraries in
order to provide some customization (i.e. using
MaterialAlertDialogBuilder with preference dialogs). This works fine,
but androidx is not happy with this and use `RestrictTo` annotations,
causing lint errors. These edits ignore this errors. It's not nice to
ignore errors, ever, but this stuff works and the errors are just
because of design decisions. Don't see how to customize stuff in other
ways, since this is the best way to customize stuff according to a
discussion I had on Android's issue tracker.